+++ lib/src/state/reading_provider.dart
@@ -2,7 +2,9 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 /// Provider for the [ReadingState].
-final readingProvider = StateNotifierProvider<ReadingNotifier, ReadingState>((ref) {
+final readingProvider = StateNotifierProvider<ReadingNotifier, ReadingState>((
+  ref,
+) {
   return ReadingNotifier();
 });